requirements.txt
Django==4.2.14
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
drf-yasg[validation]
psycopg==3.2.1
psycopg-binary==3.2.1
python-decouple==3.8
Pillow==10.4.0
django-minio-storage